change id to use the DDDD (or maybe DDDD_DDDD) number from image URL
So far this appears to be robust (every entry does have a photo) but
I've decided to let it blow up (that is, no rescue) if an ID can't be
found in the image URL. If that happens, it seems like the scraper
should fail because it will need rewriting to handle this situation
(most likely case perhaps is someone being added without a photo).

Don't know why a handful of images have the DDDD_DDDD format; couldn't
spot a useful pattern and if the pictures don't change much it doesn't
really matter.